pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agent-starter-python"
version = "1.0.0"
description = "Simple voice AI assistant built with LiveKit Agents for Python"
requires-python = ">=3.9"

dependencies = [
    "faiss-cpu>=1.13.0",
    "livekit-agents[google,openai,sarvam,silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "numpy>=2.0.2",
    "python-dotenv",
    "sentence-transformers>=5.1.2",
    "aiohttp>=3.9.0",
    "websockets>=12.0",
    "openai>=1.0.0",
    "google-cloud-texttospeech>=2.14.0",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
